requirements.txt
fastapi[all]==0.112.0
uvicorn[standard]==0.15.0
pydantic==2.8.2
pytest==7.4.4
numpy==1.21.4
transformers==4.33.3
Pillow==10.0.1
torch==2.0.1
-e git+https://github.com/toposoid/toposoid-python-lib.git@__##GIT_BRANCH##__#egg=ToposoidCommon